+++ packages/xstate-graph/src/shortestPaths.ts
@@ -17,6 +17,7 @@ import {
 } from './types';
 import { resolveTraversalOptions, createDefaultMachineOptions } from './graph';
 import { getAdjacencyMap } from './adjacency';
+import { alterPath } from './alterPath';
 
 export function getShortestPaths<
   TLogic extends AnyActorLogic,
@@ -103,7 +104,6 @@ export function getShortestPaths<
 
   weightMap.forEach(
     ({ weight, state: fromState, event: fromEvent }, stateSerial) => {
-
       const state = stateMap.get(stateSerial)!;
       const steps = !fromState
         ? []
@@ -137,34 +137,4 @@ export function getShortestPaths<
   }
 
   return paths.map(alterPath);
-}
-
-export function alterPath<T extends StatePath<any, any>>(path: T): T {
-  let steps: T['steps'] = [];
-
-  if (!path.steps.length) {
-    steps = [
-      {
-        state: path.state,
-        event: { type: 'xstate.init' } as any
-      }
-    ];
-  } else {
-    for (let i = 0; i < path.steps.length; i++) {
-      const step = path.steps[i];
-
-      steps.push({
-        state: step.state,
-        event: i === 0 ? { type: 'xstate.init' } : path.steps[i - 1].event
-      });
-    }
-    steps.push({
-      state: path.state,
-      event: path.steps[path.steps.length - 1].event
-    });
-  }
-  return {
-    ...path,
-    steps
-  };
 }
